agent/requirements.txt
tiktoken
networkx
fastapi
uvicorn
openai
sse-starlette
GitPython
PyPDF2
matplotlib
typer